ai_service/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.6.3
pandas
scikit-learn
numpy
python-dotenv
openai
groq
torch
joblib
faiss-cpu
sentence-transformers
PyMuPDF
reportlab
